修改calculateCharNum方法返回值为int; 将countCharNumByDatasetId的sql语句写在注解里

+++ ruoyi-modules/ruoyi-system/src/main/java/org/dromara/system/service/ISysOssService.java
@@ -37,5 +37,5 @@ public interface ISysOssService {
 
     Boolean deleteWithValidByIds(Collection<Long> ids, Boolean isValid);
 
-    public int calculateCharNum(Long ossId);
+    public Long calculateCharNum(Long ossId);
 }
